+++ src/config.ini
@@ -1,0 +1,14 @@
+[Models]
+ConvLSTM = {"input_dim": 4, "hidden_dim": 40, "kernel_size": [3, 3], "num_layers": 2, "physics_kernel_size": [3, 3], "output_dim": 1, "batch_first": true}
+ConvLSTM_Physics = {"input_dim": 4, "hidden_dim": 40, "kernel_size": [3, 3], "num_layers": 2, "output_dim": 1, "bias": true, "return_all_layers": false}
+ConvLSTM_Attention = {"input_dim": 4, "hidden_dim": [128, 64], "kernel_size": [3, 3], "physics_kernel_size": [3, 3], "num_layers": 2, "output_dim": 1, "batch_first": true, "bias": true, "return_all_layers": false, "window_size": 1, "num_heads": 8}
+ConvLSTM_Attention_Physics = {"input_dim": 4, "hidden_dim": [128, 64], "kernel_size": [3, 3], "physics_kernel_size": [3, 3], "num_layers": 2, "output_dim": 1, "batch_first": true, "bias": true, "return_all_layers": false, "window_size": 1, "num_heads": 8}
+
+[Schemes]
+ConvLSTM = ["standard"]
+ConvLSTM_Physics = ["physics", "physics_dynamic_grid"]
+ConvLSTM_Attention = ["standard"]
+ConvLSTM_Attention_Physics = ["physics", "physics_dynamic_grid"]
+
+[Datasets]
+radar_movies = /home/tso/RadarConvAttention/src/datasets/radar_movies.npy
